exceptions for results

+++ Student/src/main/java/com/testcreation/students/controller/ResultController.java
@@ -16,6 +16,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.testcreation.students.bean.Result;
 import com.testcreation.students.bean.Student;
 import com.testcreation.students.bean.Test;
+import com.testcreation.students.exception.StudentException;
 import com.testcreation.students.service.ResultService;
 
 
@@ -33,11 +34,15 @@ public class ResultController {
 	
 	@GetMapping("/{id}")
 	public Optional<Result> getResultById(@PathVariable Integer id){
+		if(service.getResultById(id).isEmpty())
+		throw new StudentException("Provided result id is not exists...!");
 		return service.getResultById(id);
 	}
 	
 	@GetMapping("/student/{studentId}")
-	public List<Result> getResultsByStudentId(@PathVariable Integer studentId){
+	 public List<Result> getResultsByStudentId(@PathVariable Integer studentId){
+		if(service.getResultsByStudentId(studentId).isEmpty())
+			throw new StudentException("Result doesn't exist with ID provied...!");
 		return service.getResultsByStudentId(studentId);
 	}
 	
@@ -45,16 +50,20 @@ public class ResultController {
 	void addResult(@RequestBody Result theResult,@PathVariable Integer studentId,@PathVariable Integer testId) {
 		theResult.setStudent(new Student(studentId));
 		theResult.setTest(new Test(testId));
-		service.addResult(theResult);
+		service.addResult(theResult);	
 	}
 	
 	@PutMapping("/update/{id}")
-	void updateResult(@RequestBody Result theResult) {
+	void updateResult(@RequestBody Result theResult,@PathVariable Integer id) {
+		if(service.getResultById(id).isEmpty())
+			throw new StudentException("Result doesn't exist with ID provied...!");
 		service.updateResultById(theResult);
 	}
 	
 	@DeleteMapping("/delete/{id}")
 	void deleteResult(@PathVariable Integer id) {
+		if(service.getResultById(id).isEmpty())
+			throw new StudentException("Result doesn't exist with ID provied...!");
 		service.deleteResultById(id);
 	}
